+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 Release history
 =====================================
+0.1.1
+-------------------------------------
+2014-10-11
+
+- Bootstrap 3 theme fixes: When tab pane has no or little content so
+  that the height of the dropdown menu exceeds the height of the tab pane
+  content the dropdown menu now becomes scrollable (vertically).
 
 0.1
 -------------------------------------